taking in less ram

+++ monopoly/src/board/Station.java
@@ -1,7 +1,7 @@
 package board;
 
 public class Station extends Street{
-    public Station(String naam, int plek) {
+    public Station(String naam, short plek) {
         super(20000, naam, plek, Integer.MAX_VALUE);
     }
 }
